also ingest library itself
