+++ kifi-backend/modules/graph/app/com/keepit/graph/manager/GraphUpdateFetcher.scala
@@ -54,6 +54,8 @@ class GraphUpdateFetcherImpl @Inject() (
 
       case EmailContactGraphUpdate => abook.getContactsChanged(seq.copy(), fetchSize).imap(_.map(EmailContactGraphUpdate.apply))
 
+      case LibraryGraphUpdate => shoebox.getLibrariesChanged(seq.copy(), fetchSize).imap(_.map(LibraryGraphUpdate.apply))
+
       case LibraryMembershipGraphUpdate => shoebox.getLibraryMembershipsChanged(seq.copy(), fetchSize).imap(_.map(LibraryMembershipGraphUpdate.apply))
     }
   }
